task_management_system/requirements.txt
Django==5.0.6
django-cors-headers==4.3.1
django-environ==0.11.2
django_csp==3.8
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
djoser==2.2.2
drf-nested-routers==0.94.1